pyproject.toml
[project]
name = "story-generator-2"
version = "0.1.0"
description = "A Portal to Kindred but Forgotten Histories"
requires-python = ">=3.9"
dependencies = [
    "fastapi",
    "uvicorn",
    "dspy-ai",
    "firebase-admin",
    "google-genai",
    "python-dotenv",
    "requests",
    "pydantic",
    "numpy",
    "ipdb>=0.13.13",
    "sentence-transformers>=2.2.0",
]

[tool.uv]
dev-dependencies = [
  "matplotlib>=3.9.4",
  "playwright>=1.57.0",
  "ruff>=0.8.0",
]

[tool.ruff]
line-length = 88
indent-width = 2

[tool.ruff.format]
indent-style = "space"
quote-style = "double"
